8298129: Let checkpoint event sizes grow beyond u4 limit
Backport-of: ea108f504ccb63fc9651e804e3bbba1c108dcead

+++ src/jdk.jfr/share/classes/jdk/jfr/internal/consumer/ChunkParser.java
@@ -237,7 +237,7 @@ public final class ChunkParser {
         long absoluteChunkEnd = chunkHeader.getEnd();
         while (input.position() < absoluteChunkEnd) {
             long pos = input.position();
-            int size = input.readInt();
+            long size = input.readLong();
             if (size == 0) {
                 throw new IOException("Event can't have zero size");
             }
@@ -305,7 +305,7 @@ public final class ChunkParser {
         while (thisCP != abortCP && delta != 0) {
             input.position(thisCP);
             lastCP = thisCP;
-            int size = input.readInt(); // size
+            long size = input.readLong(); // size
             long typeId = input.readLong();
             if (typeId != CONSTANT_POOL_TYPE_ID) {
                 throw new IOException("Expected check point event (id = 1) at position " + lastCP + ", but found type id = " + typeId);
